+++ sdk/src/main/java/com/hedera/hashgraph/sdk/Node.java
@@ -6,13 +6,23 @@ class Node extends ManagedNode implements Comparable<Node>{
     AccountId accountId;
     long delay;
     long delayUntil;
+    long waitTime;
 
-    Node(AccountId accountId, String address, ExecutorService executor) {
+    Node(AccountId accountId, String address, long waitTime, ExecutorService executor) {
         super(address, executor);
 
         this.accountId = accountId;
-        this.delay = 250;
+        this.delay = waitTime;
         this.delayUntil = 0;
+    }
+
+    void setWaitTime(long waitTime) {
+        // If delay is equal to the old waitTime we should change it to the new waitTime
+        if (delay == this.waitTime) {
+            delay = waitTime;
+        }
+
+        this.waitTime = waitTime;
     }
 
     boolean isHealthy() {
@@ -25,7 +35,7 @@ class Node extends ManagedNode implements Comparable<Node>{
     }
 
     void decreaseDelay() {
-        this.delay = Math.max(this.delay / 2, 250);
+        this.delay = Math.max(this.delay / 2, waitTime);
     }
 
     long delay() {